+++ src/main/java/com/maigrand/overwatchdb/payload/WeaponDetails.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class WeaponDetails {
 
     private String name;
-    private Long aimTypeId;
-    private Long heroId;
+    private Integer aimTypeId;
+    private Integer heroId;
 
 }
